requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
google-generativeai>=0.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
chardet>=5.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

